NavBar test fixes
