--- src/Menu.php
@@ -1,9 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Yiisoft\Yii\Bootstrap5;
-
-final class Menu extends AbstractNav
-{
-}
