+++ tests/NavBarTest.php
@@ -4,10 +4,17 @@ declare(strict_types=1);
 
 namespace Yiisoft\Yii\Bootstrap5\Tests;
 
+use Yiisoft\Widget\WidgetFactory;
 use Yiisoft\Yii\Bootstrap5\Collapse;
+use Yiisoft\Yii\Bootstrap5\Dropdown;
 use Yiisoft\Yii\Bootstrap5\Nav;
 use Yiisoft\Yii\Bootstrap5\NavBar;
+use Yiisoft\Yii\Bootstrap5\NavItem;
+use Yiisoft\Yii\Bootstrap5\NavLink;
 use Yiisoft\Yii\Bootstrap5\Offcanvas;
+
+use function call_user_func;
+use function is_callable;
 
 /**
  * Tests for `NavBar` widget.
@@ -102,6 +109,8 @@ final class NavBarTest extends TestCase
 
     public function testNavAndForm(): void
     {
+        $link = NavLink::widget()->url('#')->id('');
+
         $html = NavBar::widget()
             ->id('TEST_ID')
             ->collapseOptions(['id' => 'C_ID'])
@@ -110,20 +119,25 @@ final class NavBarTest extends TestCase
             ->begin();
         $html .= Nav::widget()
             ->id('N_ID')
-            ->items([
-                ['label' => 'Home', 'url' => '#'],
-                ['label' => 'Link', 'url' => '#'],
-                [
-                    'label' => 'Dropdown',
-                    'dropdownOptions' => ['id' => 'D_ID'],
-                    'items' => [
-                        ['label' => 'Action', 'url' => '#'],
-                        ['label' => 'Another action', 'url' => '#'],
-                        '-',
-                        ['label' => 'Something else here', 'url' => '#'],
-                    ],
-                ],
-            ])
+            ->links(
+                $link->label('Home'),
+                $link->label('Link'),
+                NavLink::widget()
+                    ->label('Dropdown')
+                    ->id('')
+                    ->url('#')
+                    ->item(
+                        NavItem::widget()
+                            ->dropdown(
+                                Dropdown::widget()->id('D_ID')
+                            )
+                            ->items(
+                                $link->label('Action'),
+                                $link->label('Another action'),
+                                $link->label('Something else here'),
+                            )
+                    )
+            )
             ->options(['class' => ['mr-auto']])
             ->render();
 
@@ -141,12 +155,11 @@ final class NavBarTest extends TestCase
         <a class="navbar-brand" href="/">My Company</a>
         <button type="button" class="navbar-toggler" data-bs-toggle="collapse" aria-label="Toggle navigation" aria-controls="C_ID" data-bs-target="#C_ID" aria-expanded="false"><span class="navbar-toggler-icon"></span></button>
         <div id="C_ID" class="collapse navbar-collapse">
-        <ul id="N_ID" class="mr-auto nav"><li class="nav-item"><a class="nav-link" href="#">Home</a></li>
-        <li class="nav-item"><a class="nav-link" href="#">Link</a></li>
-        <li class="dropdown nav-item"><a class="dropdown-toggle nav-link" href="#" data-bs-toggle="dropdown">Dropdown</a><ul id="D_ID" class="dropdown-menu">
+        <ul id="N_ID" class="mr-auto nav"><li class="nav-item"><a id class="nav-link" href="#">Home</a></li>
+        <li class="nav-item"><a id class="nav-link" href="#">Link</a></li>
+        <li class="nav-item dropdown"><a id class="dropdown-toggle nav-link" href="#" aria-expanded="false" role="button" data-bs-toggle="dropdown">Dropdown</a><ul id="D_ID" class="dropdown-menu">
         <li><a class="dropdown-item" href="#">Action</a></li>
         <li><a class="dropdown-item" href="#">Another action</a></li>
-        <li><hr class="dropdown-divider"></li>
         <li><a class="dropdown-item" href="#">Something else here</a></li>
         </ul></li></ul><form class="form-inline my-2 my-lg-0">
         <input class="form-control mr-sm-2" type="search" placeholder="Search" aria-label="Search">
@@ -327,6 +340,8 @@ final class NavBarTest extends TestCase
 
     public function testWithoutCollapse(): void
     {
+        $link = NavLink::widget()->url('#')->id('');
+
         $html = NavBar::widget()
             ->id('TEST_ID')
             ->brandText('My Company')
@@ -335,20 +350,25 @@ final class NavBarTest extends TestCase
             ->begin();
         $html .= Nav::widget()
             ->id('D_ID1')
-            ->items([
-                ['label' => 'Home', 'url' => '#'],
-                ['label' => 'Link', 'url' => '#'],
-                [
-                    'label' => 'Dropdown',
-                    'dropdownOptions' => ['id' => 'D_ID2'],
-                    'items' => [
-                        ['label' => 'Action', 'url' => '#'],
-                        ['label' => 'Another action', 'url' => '#'],
-                        '-',
-                        ['label' => 'Something else here', 'url' => '#'],
-                    ],
-                ],
-            ])
+            ->links(
+                $link->label('Home'),
+                $link->label('Link'),
+                NavLink::widget()
+                    ->id('')
+                    ->label('Dropdown')
+                    ->url('#')
+                    ->item(
+                        NavItem::widget()
+                            ->dropdown(
+                                Dropdown::widget()->id('D_ID2')
+                            )
+                            ->items(
+                                $link->label('Action'),
+                                $link->label('Another action'),
+                                $link->label('Something else here'),
+                            )
+                    )
+            )
             ->options(['class' => ['mr-auto']])
             ->render();
         $html .= NavBar::end();
@@ -357,12 +377,11 @@ final class NavBarTest extends TestCase
         <nav id="TEST_ID" class="navbar">
         <div class="container">
         <a class="navbar-brand" href="/">My Company</a>
-        <ul id="D_ID1" class="mr-auto nav"><li class="nav-item"><a class="nav-link" href="#">Home</a></li>
-        <li class="nav-item"><a class="nav-link" href="#">Link</a></li>
-        <li class="dropdown nav-item"><a class="dropdown-toggle nav-link" href="#" data-bs-toggle="dropdown">Dropdown</a><ul id="D_ID2" class="dropdown-menu">
+        <ul id="D_ID1" class="mr-auto nav"><li class="nav-item"><a id class="nav-link" href="#">Home</a></li>
+        <li class="nav-item"><a id class="nav-link" href="#">Link</a></li>
+        <li class="nav-item dropdown"><a id class="dropdown-toggle nav-link" href="#" aria-expanded="false" role="button" data-bs-toggle="dropdown">Dropdown</a><ul id="D_ID2" class="dropdown-menu">
         <li><a class="dropdown-item" href="#">Action</a></li>
         <li><a class="dropdown-item" href="#">Another action</a></li>
-        <li><hr class="dropdown-divider"></li>
         <li><a class="dropdown-item" href="#">Something else here</a></li>
         </ul></li></ul></div>
         </nav>
@@ -611,12 +630,12 @@ final class NavBarTest extends TestCase
     {
         return [
             [
-                Collapse::widget()->id('T_ID'),
+                static fn () => Collapse::widget()->id('T_ID'),
                 '<button type="button" data-bs-toggle="collapse" aria-controls="T_ID" data-bs-target="#T_ID" aria-expanded="false"></button>',
             ],
 
             [
-                Offcanvas::widget()->id('T_ID'),
+                static fn () => Offcanvas::widget()->id('T_ID'),
                 '<button type="button" data-bs-toggle="offcanvas" aria-controls="T_ID" data-bs-target="#T_ID"></button>',
             ],
 
@@ -635,8 +654,12 @@ final class NavBarTest extends TestCase
     /**
      * @dataProvider toggleWidgetDataProvider
      */
-    public function testToggle(Collapse|Offcanvas|null $widget, string $expected): void
+    public function testToggle(?callable $widget, string $expected): void
     {
+        if (is_callable($widget)) {
+            $widget = call_user_func($widget);
+        }
+
         $navBar = NavBar::widget()
             ->id('TEST_ID')
             ->withWidget($widget);
